+++ src/beat_link_trigger/show_cues.clj
@@ -311,46 +311,53 @@
   [(seesaw/action :name "Entered"
                   :enabled? (cue-event-enabled? context cue :entered)
                   :handler (fn [_]
-                             (let [[_show context runtime-info] (latest-show-and-context context)]
-                               (send-cue-messages context runtime-info cue :entered
-                                                  (random-status-for-simulation :entered context)))))
+                             (binding [util/*simulating* true]
+                               (let [[_show context runtime-info] (latest-show-and-context context)]
+                                 (send-cue-messages context runtime-info cue :entered
+                                                    (random-status-for-simulation :entered context))))))
    (seesaw/action :name "Started On-Beat"
                   :enabled? (cue-event-enabled? context cue :started-on-beat)
                   :handler (fn [_]
-                             (let [[_show context runtime-info] (latest-show-and-context context)]
-                               (send-cue-messages context runtime-info cue :started-on-beat
-                                                  (random-status-for-simulation :started-on-beat context)))))
+                             (binding [util/*simulating* true]
+                               (let [[_show context runtime-info] (latest-show-and-context context)]
+                                 (send-cue-messages context runtime-info cue :started-on-beat
+                                                    (random-status-for-simulation :started-on-beat context))))))
    (seesaw/action :name "Started Late"
                   :enabled? (cue-event-enabled? context cue :started-late)
                   :handler (fn [_]
-                             (let [[_show context runtime-info] (latest-show-and-context context)]
-                               (send-cue-messages context runtime-info cue :started-late
-                                                  (random-status-for-simulation :started-late context)))))
+                             (binding [util/*simulating* true]
+                               (let [[_show context runtime-info] (latest-show-and-context context)]
+                                 (send-cue-messages context runtime-info cue :started-late
+                                                    (random-status-for-simulation :started-late context))))))
    (seesaw/action :name "Beat"
                   :enabled? (not (cue-missing-expression? context cue :beat))
-                  :handler (fn [_] (run-cue-function context cue :beat
-                                                     (random-status-for-simulation :beat context)
-                                                     true)))
+                  :handler (fn [_] (binding [util/*simulating* true]
+                                     (run-cue-function context cue :beat
+                                                       (random-status-for-simulation :beat context)
+                                                       true))))
    (seesaw/action :name "Tracked Update"
                   :enabled? (not (cue-missing-expression? context cue :tracked))
-                  :handler (fn [_] (run-cue-function context cue :tracked
-                                                     (random-status-for-simulation :tracked context)
-                                                     true)))
+                  :handler (fn [_] (binding [util/*simulating* true]
+                                     (run-cue-function context cue :tracked
+                                                       (random-status-for-simulation :tracked context)
+                                                       true))))
    (let [enabled-events (filterv (partial cue-event-enabled? context cue) [:started-on-beat :started-late])]
      (seesaw/action :name "Ended"
                     :enabled? (seq enabled-events)
                     :handler (fn [_]
                                (su/swap-context-runtime! nil context assoc-in [:cues (:uuid cue) :last-entry-event]
                                                          (rand-nth enabled-events))
-                               (let [[_show context runtime-info] (latest-show-and-context context)]
-                                 (send-cue-messages context runtime-info cue :ended
-                                                    (random-status-for-simulation :ended context))))))
+                               (binding [util/*simulating* true]
+                                 (let [[_show context runtime-info] (latest-show-and-context context)]
+                                   (send-cue-messages context runtime-info cue :ended
+                                                      (random-status-for-simulation :ended context)))))))
    (seesaw/action :name "Exited"
                   :enabled? (cue-event-enabled? context cue :entered)
                   :handler (fn [_]
-                             (let [[_show context runtime-info] (latest-show-and-context context)]
-                               (send-cue-messages context runtime-info cue :exited
-                                                  (random-status-for-simulation :exited context)))))])
+                             (binding [util/*simulating* true]
+                               (let [[_show context runtime-info] (latest-show-and-context context)]
+                                 (send-cue-messages context runtime-info cue :exited
+                                                    (random-status-for-simulation :exited context))))))])
 
 (defn- cue-simulate-menu
   "Creates the submenu containing actions that simulate events happening